Move test to cargo_command.rs
